fixed install

+++ run.sh
@@ -1,13 +1,27 @@
 #!/bin/bash
 
-source env/bin/activate
+prg=$0
+if [ ! -e "$prg" ]; then
+  case $prg in
+    (*/*) exit 1;;
+    (*) prg=$(command -v -- "$prg") || exit;;
+  esac
+fi
+dir=$(
+  cd -P -- "$(dirname -- "$prg")" && pwd -P
+) || exit
+prg="$dir"/launch.py
+
+source "$dir"/env/bin/activate
 echo "launching airshare widget..."
-python launch.py
+python $prg
 deactivate
 
 PID=`pgrep launch.py`
-kill -9 $PID
+if [ ! -z "$PID" ]; then
+  kill -9 $PID
+fi
 
-if [ -f tmp/temp.png ]; then
-  rm tmp/temp.png
+if [ -f "$dir"/tmp/temp.png ]; then
+  rm "$dir"/tmp/temp.png
 fi
